fix: deprecate bearer strategy
