+++ .golangci.yml
@@ -48,7 +48,3 @@ issues:
         - errcheck
         - gosec
 
-    # FIXME temporarily suppress this.
-    - path: "auth/strategies/(token|bearer)/"
-      linters:
-        - dupl
