--- auth/strategies/bearer/testdata/invalid_columns.csv
@@ -1,2 +1,0 @@
-testUserToken,testUser,1,"group1"
-testUserToken3,testUser3
