--- auth/strategies/bearer/testdata/invalid_token.csv
@@ -1,3 +1,0 @@
-testUserToken,testUser,1,"group1"
-,testUser2,2,"group1,group2","extension=1,example=2"
-testUserToken3,testUser3,3
